锚点拖动时增加边界判断

+++ smartcropperlib/src/main/java/me/pqpo/smartcropperlib/utils/CropUtils.java
@@ -12,7 +12,7 @@ public class CropUtils {
         return getPointsDistance(p1.x, p1.y, p2.x, p2.y);
     }
 
-    public static double getPointsDistance(int x1, int y1, int x2, int y2) {
+    public static double getPointsDistance(float x1, float y1, float x2, float y2) {
         return Math.sqrt(Math.pow(x1 - x2, 2) + Math.pow(y1 - y2, 2));
     }
